+++ src/main/js/pane/plain-tweakpane.ts
@@ -1,22 +1,22 @@
-import {RootApi} from './api/root';
-import {RootController} from './controller/root';
-import {ClassName} from './misc/class-name';
-import * as DomUtil from './misc/dom-util';
-import {PaneError} from './misc/pane-error';
-import {TypeUtil} from './misc/type-util';
-import {ViewModel} from './model/view-model';
-import {InputBindingPlugin} from './plugin/input-binding';
-import {BooleanInputPlugin} from './plugin/input-bindings/boolean';
-import {NumberColorInputPlugin} from './plugin/input-bindings/color-number';
-import {ObjectColorInputPlugin} from './plugin/input-bindings/color-object';
-import {StringColorInputPlugin} from './plugin/input-bindings/color-string';
-import {NumberInputPlugin} from './plugin/input-bindings/number';
-import {Point2dInputPlugin} from './plugin/input-bindings/point-2d';
-import {StringInputPlugin} from './plugin/input-bindings/string';
-import {MonitorBindingPlugin} from './plugin/monitor-binding';
-import {BooleanMonitorPlugin} from './plugin/monitor-bindings/boolean';
-import {NumberMonitorPlugin} from './plugin/monitor-bindings/number';
-import {StringMonitorPlugin} from './plugin/monitor-bindings/string';
+import {RootApi} from '../api/root';
+import {RootController} from '../controller/root';
+import {ClassName} from '../misc/class-name';
+import * as DomUtil from '../misc/dom-util';
+import {PaneError} from '../misc/pane-error';
+import {TypeUtil} from '../misc/type-util';
+import {ViewModel} from '../model/view-model';
+import {InputBindingPlugin} from '../plugin/input-binding';
+import {BooleanInputPlugin} from '../plugin/input-bindings/boolean';
+import {NumberColorInputPlugin} from '../plugin/input-bindings/color-number';
+import {ObjectColorInputPlugin} from '../plugin/input-bindings/color-object';
+import {StringColorInputPlugin} from '../plugin/input-bindings/color-string';
+import {NumberInputPlugin} from '../plugin/input-bindings/number';
+import {Point2dInputPlugin} from '../plugin/input-bindings/point-2d';
+import {StringInputPlugin} from '../plugin/input-bindings/string';
+import {MonitorBindingPlugin} from '../plugin/monitor-binding';
+import {BooleanMonitorPlugin} from '../plugin/monitor-bindings/boolean';
+import {NumberMonitorPlugin} from '../plugin/monitor-bindings/number';
+import {StringMonitorPlugin} from '../plugin/monitor-bindings/string';
 import {TweakpaneConfig} from './tweakpane-config';
 
 function createDefaultWrapperElement(document: Document): HTMLElement {
